requirements.txt
Flask==2.0.1
gitpython==3.1.14
python-dotenv==0.17.1
